// frontend/src/app/api-client/model/exchangeRateResponse.ts
/**
 * Exchange Rate Management System API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExchangeRateResponse { 
    fromCurrency: string;
    toCurrency: string;
    /**
     * Spread-adjusted rate, exact decimal (transported as string to avoid float precision loss)
     */
    rate: string;
    rateDate: string;
    fromCurrencyUsageCount: number;
    toCurrencyUsageCount: number;
}